refactor: Config에서 Manager로 클래스명 변경, controller run메서드명 변경
